added new permissions file in icds reports
